requirements.txt
beautifulsoup4
matplotlib==3.3.*
numpy==1.19.*
pandas==1.1*
Pillow==8.0.*
pip==20.2.*
requests==2.24*
selenium==3.141.*
ipykernel
jupyterlab
python-dotenv